+++ src/unix/notbsd/linux/musl/b32/arm.rs
@@ -303,4 +303,4 @@ pub const TIOCMSET: ::c_ulong = 0x5418;
 pub const FIONREAD: ::c_ulong = 0x541B;
 pub const TIOCCONS: ::c_ulong = 0x541D;
 
-pub const SYS_gettid: ::c_int = 224;
+pub const SYS_gettid: ::c_long = 224;
