+++ src/unix/notbsd/linux/other/b64/x86_64.rs
@@ -95,7 +95,7 @@ pub const PTRACE_SETFPXREGS: ::c_uint = 19;
 pub const PTRACE_GETREGS: ::c_uint = 12;
 pub const PTRACE_SETREGS: ::c_uint = 13;
 
-pub const SYS_gettid: ::c_int = 186;
+pub const SYS_gettid: ::c_long = 186;
 
 extern {
     pub fn getcontext(ucp: *mut ucontext_t) -> ::c_int;
